+++ daaukins/server/store/store.go
@@ -23,7 +23,7 @@ type ChallengeTemplate struct {
 	Name   string `yaml:"name"`
 	Id     string `yaml:"id"`
 	Image  string `yaml:"image"`
-	Memory uint64 `yaml:"memory"`
+	Memory int    `yaml:"memory"`
 }
 
 type Store struct {
