+++ daaukins/server/utils/utils_test.go
@@ -48,7 +48,7 @@ func TestGetAvailableMemoryHasZeroMemory(t *testing.T) {
 	}
 }
 
-func createMockMeminfoFile(t *testing.T, memAvailableMb uint64) (func(), error) {
+func createMockMeminfoFile(t *testing.T, memAvailableMb int) (func(), error) {
 	meminfoContents := fmt.Sprintf(`MemTotal:       16270828 kB
 MemFree:         3745292 kB
 MemAvailable:    %d kB
